requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
python-multipart
twilio
requests
python-dotenv
httpx
huggingface_hub
streamlit